requirements.txt
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.3.1
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.17.0
fire==0.7.0
flatbuffers==25.1.24
fsspec==2025.2.0
future==1.0.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.3
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sympy==1.13.3
termcolor==2.5.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
